requirements.txt
flask
flask-cors
openai
scikit-learn
joblib
python-dotenv
pandas